Moved everything over to Python3 and a re-architecting.
Moved the application over to Python3 and split the program into
multiple modules to faciliate adding new features. In particular,
support for different schedulers for deciding which node to use
and adding a logger to the central system logger. Also cleaned up
a lot of different config file editors.

As of this commit, remote transcoder is testing and working.